fix(limits): draftErrors moves to a dependency-free module (hermetic suite)
CI runs the helper suite with NO node_modules; importing draftErrors from
useConfigDraft.js pulled react through instanceNames.js and broke the
hermetic property (caught by 149's merge-ref CI). The validation seam now
lives in draftErrors.js with zero imports; the name-rule constants move
with it (instanceNames.js re-exports for the react side) and contracts.mjs
greps the new home. Verified hermetically: helpers green with node_modules
removed.
