+++ admin/spa/tests/contracts.mjs
@@ -27,19 +27,22 @@ const check = (name, fn) => {
   }
 };
 
+// both mirror regexes live in draftErrors.js since the hermetic split
+// (2026-08-13): the validation seam must stay importable with no
+// node_modules, so the literals moved to the dependency-free module
 check("instance-name regex matches config._INSTANCE_NAME_RE", () => {
-  const src = readFileSync(join(here, "../src/lib/instanceNames.js"), "utf8");
+  const src = readFileSync(join(here, "../src/lib/draftErrors.js"), "utf8");
   assert.ok(src.includes(`/${contracts.instance_name_re}/`),
-    "instanceNames.js INSTANCE_NAME_RE drifted from the contract");
+    "draftErrors.js INSTANCE_NAME_RE drifted from the contract");
   const re = new RegExp(`^${contracts.instance_name_re}$`);
   assert.ok(re.test("linear"), "accepts a valid instance name");
   assert.ok(!re.test("Bad-Name"), "rejects uppercase / hyphen");
 });
 
 check("harness-var regex matches config.HARNESS_VAR_PATTERN", () => {
-  const src = readFileSync(join(here, "../src/lib/useConfigDraft.js"), "utf8");
+  const src = readFileSync(join(here, "../src/lib/draftErrors.js"), "utf8");
   assert.ok(src.includes(`/^${contracts.harness_var_pattern}$/`),
-    "useConfigDraft's inline harness-var regex drifted from the contract");
+    "draftErrors' inline harness-var regex drifted from the contract");
   const re = new RegExp(`^${contracts.harness_var_pattern}$`);
   assert.ok(re.test("ANTHROPIC_API_KEY"), "accepts a harness var");
   assert.ok(!re.test("not-a-var"), "rejects lowercase / hyphen");
